refactor: prune and move FunctionalConnectivityParcels
