--- junifer/markers/functional_connectivity_parcels.py
@@ -1,153 +1,0 @@
-"""Provide class for functional connectivity."""
-
-# Authors: Amir Omidvarnia <a.omidvarnia@fz-juelich.de>
-#          Kaustubh R. Patil <k.patil@fz-juelich.de>
-# License: AGPL
-
-from typing import Any, Dict, List, Optional, Union
-
-from nilearn.connectome import ConnectivityMeasure
-from sklearn.covariance import EmpiricalCovariance
-
-from ..api.decorators import register_marker
-from .base import BaseMarker
-from .parcel_aggregation import ParcelAggregation
-
-
-@register_marker
-class FunctionalConnectivityParcels(BaseMarker):
-    """Class for functional connectivity.
-
-    Parameters
-    ----------
-    parcellation : str or list of str
-        The name(s) of the parcellation(s). Check valid options by calling
-        :func:`junifer.data.parcellations.list_parcellations`.
-    agg_method : str, optional
-        The method to perform aggregation using. Check valid options in
-        :func:`junifer.stats.get_aggfunc_by_name` (default "mean").
-    agg_method_params : dict, optional
-        Parameters to pass to the aggregation function. Check valid options in
-        :func:`junifer.stats.get_aggfunc_by_name` (default None).
-    cor_method : str, optional
-        The method to perform correlation using. Check valid options in
-        :class:`nilearn.connectome.ConnectivityMeasure`
-        (default "covariance").
-    cor_method_params : dict, optional
-        Parameters to pass to the correlation function. Check valid options in
-        :class:`nilearn.connectome.ConnectivityMeasure` (default None).
-    mask : str, optional
-        The name of the mask to apply to regions before extracting signals.
-        Check valid options by calling :func:`junifer.data.masks.list_masks`
-        (default None).
-    name : str, optional
-        The name of the marker. If None, will use the class name (default
-        None).
-    """
-
-    _DEPENDENCIES = {"nilearn", "scikit-learn"}
-
-    def __init__(
-        self,
-        parcellation: Union[str, List[str]],
-        agg_method: str = "mean",
-        agg_method_params: Optional[Dict] = None,
-        cor_method: str = "covariance",
-        cor_method_params: Optional[Dict] = None,
-        mask: Optional[str] = None,
-        name: Optional[str] = None,
-    ) -> None:
-        self.parcellation = parcellation
-        self.agg_method = agg_method
-        self.agg_method_params = agg_method_params
-        self.cor_method = cor_method
-        self.cor_method_params = cor_method_params or {}
-
-        # default to nilearn behavior
-        self.cor_method_params["empirical"] = self.cor_method_params.get(
-            "empirical", False
-        )
-        self.mask = mask
-
-        super().__init__(name=name)
-
-    def get_valid_inputs(self) -> List[str]:
-        """Get valid data types for input.
-
-        Returns
-        -------
-        list of str
-            The list of data types that can be used as input for this marker.
-        """
-        return ["BOLD"]
-
-    def get_output_type(self, input_type: str) -> str:
-        """Get output type.
-
-        Parameters
-        ----------
-        input_type : str
-            The data type input to the marker.
-
-        Returns
-        -------
-        str
-            The storage type output by the marker.
-
-        """
-        return "matrix"
-
-    def compute(
-        self,
-        input: Dict[str, Any],
-        extra_input: Optional[Dict] = None,
-    ) -> Dict:
-        """Compute.
-
-        Parameters
-        ----------
-        input : dict
-            A single input from the pipeline data object in which to compute
-            the marker.
-        extra_input : dict, optional
-            The other fields in the pipeline data object. Useful for accessing
-            other data kind that needs to be used in the computation. For
-            example, the functional connectivity markers can make use of the
-            confounds if available (default None).
-
-        Returns
-        -------
-        dict
-            The computed result as dictionary. The following data will be
-            included in the dictionary:
-
-            * ``data`` : functional connectivity matrix as a numpy.ndarray.
-            * ``row_names`` : row names as a list
-            * ``col_names`` : column names as a list
-            * ``matrix_kind`` : the kind of matrix (tril, triu or full)
-
-        """
-        pa = ParcelAggregation(
-            parcellation=self.parcellation,
-            method=self.agg_method,
-            method_params=self.agg_method_params,
-            mask=self.mask,
-            on="BOLD",
-        )
-        # get the 2D timeseries after parcel aggregation
-        ts = pa.compute(input)
-
-        if self.cor_method_params["empirical"]:
-            cm = ConnectivityMeasure(
-                cov_estimator=EmpiricalCovariance(),  # type: ignore
-                kind=self.cor_method,
-            )
-        else:
-            cm = ConnectivityMeasure(kind=self.cor_method)
-        out = {}
-        out["data"] = cm.fit_transform([ts["data"]])[0]
-        # create column names
-        out["row_names"] = ts["columns"]
-        out["col_names"] = ts["columns"]
-        out["matrix_kind"] = "tril"
-        return out
